requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.6.0